snapper/requirements.txt
selenium
selenium-screenshot
pillow
nose
flask
virtualenv